build: 📦️ 减少不必要的依赖
