+++ package.json
@@ -1,11 +1,11 @@
 {
-  "name": "truth-test-cli",
-  "version": "1.0.1",
+  "name": "truth-cli",
+  "version": "1.0.2",
   "description": "A command-line tool for analyzing dependencies under node_moudles",
   "type": "module",
   "main": "dist/index.js",
   "bin": {
-    "truth-test-cli": "dist/index.js"
+    "truth-cli": "dist/index.js"
   },
   "scripts": {
     "init": "pnpm install",
@@ -36,9 +36,7 @@
   },
   "homepage": "https://github.com/lilming123/dep_cil#readme",
   "dependencies": {
-    "@types/node": "^20.4.5",
     "commander": "^11.0.0",
-    "execa": "^7.2.0",
     "express": "^4.18.2"
   },
   "devDependencies": {
@@ -52,15 +50,10 @@
     "@rollup/plugin-typescript": "^11.1.2",
     "@types/express": "^4.17.17",
     "@types/minimist": "^1.2.2",
-    "@types/node": "^20.4.5",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
     "chalk": "^5.3.0",
     "commander": "^11.0.0",
     "commitizen": "^4.3.0",
     "cz-git": "^1.7.0",
-    "esbuild": "^0.18.17",
-    "esbuild-plugin-polyfill-node": "^0.3.0",
     "eslint": "^8.45.0",
     "husky": "^8.0.3",
     "lint-staged": "^13.2.3",
@@ -71,9 +64,7 @@
     "rollup": "^3.26.3",
     "rollup-plugin-polyfill-node": "^0.12.0",
     "ts-node": "^10.9.1",
-    "tslib": "^2.6.1",
-    "typescript": "^5.1.6",
-    "vite": "^4.4.5"
+    "typescript": "^5.1.6"
   },
   "config": {
     "commitizen": {
