+++ packages/web/.eslintrc.cjs
@@ -1,10 +1,3 @@
 module.exports = {
   extends: '@antfu',
-  parserOptions: {
-    ecmaVersion: 'latest',
-    parser: '@typescript-eslint/parser',
-    sourceType: 'module',
-  },
-  plugins: ['@typescript-eslint', 'vue'],
-  rules: {},
 }
