+++ packages/web/package.json
@@ -17,11 +17,8 @@
   "devDependencies": {
     "@antfu/eslint-config": "^0.39.8",
     "@types/echarts": "^4.9.18",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
     "@vitejs/plugin-vue": "^4.2.3",
     "eslint": "^8.45.0",
-    "eslint-plugin-vue": "^9.15.1",
     "typescript": "^5.0.2",
     "vite": "^4.4.5",
     "vue-tsc": "^1.8.5"
